# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Vaccination API
  description: This doc describe all the features and api-end points of vaccination app
  contact:
    name: Dev
    email: dev@dev.com
    url: 'https://dev.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: Dev
    url: 'http://localhost:8000/api/v1'
  - description: SwaggerHub API Auto Mocking
    url: 
      https://virtserver.swaggerhub.com/Learning-SwaggerAPI/Vaccination-API/1.0.0
...
paths:
...
  '/appointments/{id}':
    get:
      operationId: getSingleAppointment
      security:
        - bearerAuth: []
      tags:
        - Appointments
      description: Retrive a single appointment
      parameters:
        - $ref: '#/components/parameters/idPathParam'
        - in: query
          name: expand
          description: expand reviews
          schema:
            type: string
            example: "user"
      responses:
        '200':
          description: Appoinment details fetched successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Appointment'
                      - type: object
                        properties:
                          reviews:
                            type: array
                            items:
                              $ref: '#/components/schemas/Reviews'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /appointments/1
                      appointments:
                        type: string
                        example: /appointments
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    patch:
      security:
        - bearerAuth: []
      tags:
        - Appointments
      description: Partially update an appointment
      parameters:
        - $ref: '#/components/parameters/idPathParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                vaccine:
                  type: string
                  example: Pnemonia shot
                date:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Successfully Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentUpdateDTO'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      security:
        - bearerAuth: []
      tags:
        - Appointments
      description: Delete an appointment by id
      parameters:
        - $ref: '#/components/parameters/idPathParam'
      responses:
        '204':
          description: Appointment deleted successfully
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Appointment:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
            userId:
              type: integer
              format: int32
              example: 1
            vaccine:
              type: string
              example: Covisheild
            status:
              type: string
              enum:
                - Complete
                - Pending
              example: Complete
            appintment_at:
              format: date-time
          required:
            - id
            - vaccine
            - appintment_at
        - $ref: '#/components/schemas/Timestamp'
...
    AppointmentUpdateDTO:
      type: object
      properties:
        message:
          type: string
          example: Appointment updated successfully
        data:
          $ref: '#/components/schemas/Appointment'
        links:
          type: object
          properties:
            self:
              type: string
              example: /appointments/1
            appoinments:
              type: string
              example: /appointments
    Reviews:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 1
            userId:
              type: integer
              format: int32
              example: 1
            vaccineId:
              type: integer
              format: int32
              example: 1
            comment:
              type: string
              format: text
              example: The body of the comment
            status:
              type: string
              enum:
                - public
                - private
              example: public
          required:
            - id
            - userId
            - vaccineId
            - body
            - status
        - $ref: '#/components/schemas/Timestamp'
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - createdAt
        - updatedAt
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Bad Request
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    '404':
      description: Resource Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Requested resource not found
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal Server Error
...
  parameters:
...
    idPathParam:
      in: path
      name: id
      required: true
      description: item Id
      schema:
        type: string
        example: 1
...